Make sure we don't modify the body of a message when injecting a header.
When we come across a part that has no specific content-type, we inject one
to satisfy RFC 1341, section 7.2. We must inject a header because Email::MIME
body parts are just Email::Simple objects -- which expect messages that have at
least one header.

After we've generated the part, we can then remove the added header, and be
on our merry way. This is awful, but works.