do not allow an excessive number of parts
This is annoying!  It bugs me that making a mere 100,000 parts should lead
to excessive consumption of memory… but really, nobody does that for
a good reason.

This makes top-level (non-recursive) Email::MIME->new set a localized
part counter to 0 and increment before it makes new subpart objects.
If the counter exceeds the max, it will die early.

It would be better to work, but this is better than eating all memory and
crashing, right?

Put together with help from Marc Bradshaw.

Test forthcoming.